test: split AgentFS file contract suite
